#572 created relation-type-store to break the tight coupling between relation, hasone, and hasmany. moved store and relation into their own module
