+++ src/backbone-relational.js
@@ -7,6 +7,10 @@ import eventQueue from './event-queue';
 import BObject from './utils/object';
 import config from './config';
 import Collection from './collection';
+import relationTypeStore from './relation-type-store';
+import Store from './utils/store';
+import store from './store';
+import Relation from './relation';
 
 const module = config;
 
@@ -14,623 +18,12 @@ module.Collection = Collection;
 module.Semaphore = Semaphore;
 module.BlockingQueue = BlockingQueue;
 module.eventQueue = eventQueue;
+module.relationTypeStore = relationTypeStore;
 
-/**
- * Backbone.Store keeps track of all created (and destruction of) Backbone.Relational.Model.
- * Handles lookup for relations.
- */
-module.Store = BObject.extend({
-	initialize() {
-		this._collections = [];
-		this._reverseRelations = [];
-		this._orphanRelations = [];
-		this._subModels = [];
-		this._modelScopes = [ window ];
-	},
+module.Store = Store;
+module.store = store;
 
-	/**
-	 * Create a new `Relation`.
-	 * @param {Backbone.Relational.Model} [model]
-	 * @param {Object} relation
-	 * @param {Object} [options]
-	 */
-	initializeRelation: function( model, relation, options ) {
-		let { type: Type } = relation;
-		if ( _.isString( Type ) ) {
-			Type = module[ Type ] || this.getObjectByName( Type );
-		}
-
-		if ( _.isObject( Type ) ) {
-			let relationType = new Type( model, relation, options ); // Also pushes the new Relation into `model._relations`
-		}
-		else if ( config.showWarnings && console ) {
-			console.warn( 'Relation=%o; missing or invalid relation type!', relation );
-		}
-	},
-
-	/**
-	 * Add a scope for `getObjectByName` to look for model types by name.
-	 * @param {Object} scope
-	 */
-	addModelScope: function( scope ) {
-		this._modelScopes.push( scope );
-	},
-
-	/**
-	 * Remove a scope.
-	 * @param {Object} scope
-	 */
-	removeModelScope: function( scope ) {
-		this._modelScopes = _.without( this._modelScopes, scope );
-	},
-
-	/**
-	 * Add a set of subModelTypes to the store, that can be used to resolve the '_superModel'
-	 * for a model later in 'setupSuperModel'.
-	 *
-	 * @param {Backbone.Relational.Model} subModelTypes
-	 * @param {Backbone.Relational.Model} superModelType
-	 */
-	addSubModels: function( subModelTypes, superModelType ) {
-		this._subModels.push({
-			superModelType,
-			subModels: subModelTypes
-		});
-	},
-
-	/**
-	 * Check if the given modelType is registered as another model's subModel. If so, add it to the super model's
-	 * '_subModels', and set the modelType's '_superModel', '_subModelTypeName', and '_subModelTypeAttribute'.
-	 *
-	 * @param {Backbone.Relational.Model} modelType
-	 */
-	setupSuperModel: function( modelType ) {
-		_.find( this._subModels, function( subModelDef ) {
-			return _.filter( subModelDef.subModels || [], function( subModelTypeName, typeValue ) {
-				var subModelType = this.getObjectByName( subModelTypeName );
-
-				if ( modelType === subModelType ) {
-					// Set 'modelType' as a child of the found superModel
-					subModelDef.superModelType._subModels[ typeValue ] = modelType;
-
-					// Set '_superModel', '_subModelTypeValue', and '_subModelTypeAttribute' on 'modelType'.
-					modelType._superModel = subModelDef.superModelType;
-					modelType._subModelTypeValue = typeValue;
-					modelType._subModelTypeAttribute = subModelDef.superModelType.prototype.subModelTypeAttribute;
-					return true;
-				}
-			}, this ).length;
-		}, this );
-	},
-
-	/**
-	 * Add a reverse relation. Is added to the 'relations' property on model's prototype, and to
-	 * existing instances of 'model' in the store as well.
-	 * @param {Object} relation
-	 * @param {Backbone.Relational.Model} relation.model
-	 * @param {String} relation.type
-	 * @param {String} relation.key
-	 * @param {String|Object} relation.relatedModel
-	 */
-	addReverseRelation: function( relation ) {
-		var exists = _.any( this._reverseRelations, function( rel ) {
-			return _.all( relation || [], function( val, key ) {
-				return val === rel[ key ];
-			});
-		});
-
-		if ( !exists && relation.model && relation.type ) {
-			this._reverseRelations.push( relation );
-			this._addRelation( relation.model, relation );
-			this.retroFitRelation( relation );
-		}
-	},
-
-	/**
-	 * Deposit a `relation` for which the `relatedModel` can't be resolved at the moment.
-	 *
-	 * @param {Object} relation
-	 */
-	addOrphanRelation: function( relation ) {
-		var exists = _.any( this._orphanRelations, function( rel ) {
-			return _.all( relation || [], function( val, key ) {
-				return val === rel[ key ];
-			});
-		});
-
-		if ( !exists && relation.model && relation.type ) {
-			this._orphanRelations.push( relation );
-		}
-	},
-
-	/**
-	 * Try to initialize any `_orphanRelation`s
-	 */
-	processOrphanRelations: function() {
-		// Make sure to operate on a copy since we're removing while iterating
-		_.each( this._orphanRelations.slice( 0 ), function( rel ) {
-			var relatedModel = this.getObjectByName( rel.relatedModel );
-			if ( relatedModel ) {
-				this.initializeRelation( null, rel );
-				this._orphanRelations = _.without( this._orphanRelations, rel );
-			}
-		}, this );
-	},
-
-	/**
-	 *
-	 * @param {Backbone.Relational.Model.constructor} type
-	 * @param {Object} relation
-	 * @private
-	 */
-	_addRelation: function( type, relation ) {
-		if ( !type.prototype.relations ) {
-			type.prototype.relations = [];
-		}
-		type.prototype.relations.push( relation );
-
-		_.each( type._subModels || [], function( subModel ) {
-			this._addRelation( subModel, relation );
-		}, this );
-	},
-
-	/**
-	 * Add a 'relation' to all existing instances of 'relation.model' in the store
-	 * @param {Object} relation
-	 */
-	retroFitRelation: function( relation ) {
-		let { type: RelationType } = relation;
-
-		var coll = this.getCollection( relation.model, false );
-		coll && coll.each( function( model ) {
-			if ( !( model instanceof relation.model ) ) {
-				return;
-			}
-
-			let relationType = new RelationType( model, relation );
-		}, this );
-	},
-
-	/**
-	 * Find the Store's collection for a certain type of model.
-	 * @param {Backbone.Relational.Model} type
-	 * @param {Boolean} [create=true] Should a collection be created if none is found?
-	 * @return {module.Collection} A collection if found (or applicable for 'model'), or null
-	 */
-	getCollection: function( type, create ) {
-		if ( type instanceof Backbone.Model ) {
-			type = type.constructor;
-		}
-
-		var rootModel = type;
-		while ( rootModel._superModel ) {
-			rootModel = rootModel._superModel;
-		}
-
-		var coll = _.find( this._collections, function( item ) {
-			return item.model === rootModel;
-		});
-
-		if ( !coll && create !== false ) {
-			coll = this._createCollection( rootModel );
-		}
-
-		return coll;
-	},
-
-	/**
-	 * Find a model type on one of the modelScopes by name. Names are split on dots.
-	 * @param {String} name
-	 * @return {Object}
-	 */
-	getObjectByName: function( name ) {
-		var parts = name.split( '.' ),
-			type = null;
-
-		_.find( this._modelScopes, function( scope ) {
-			type = _.reduce( parts || [], function( memo, val ) {
-				return memo ? memo[ val ] : undefined;
-			}, scope );
-
-			if ( type && type !== scope ) {
-				return true;
-			}
-		}, this );
-
-		return type;
-	},
-
-	_createCollection: function( type ) {
-		// If 'type' is an instance, take its constructor
-		if ( !_.isObject( type ) ) {
-			type = type.constructor;
-		}
-
-		// Type should inherit from Backbone.Relational.Model.
-		// if ( type.prototype instanceof module.Model ) {
-		let coll = new module.Collection();
-		coll.model = type;
-
-		this._collections.push( coll );
-		// }
-
-		return coll;
-	},
-
-	/**
-	 * Find the attribute that is to be used as the `id` on a given object
-	 * @param type
-	 * @param {String|Number|Object|Backbone.Relational.Model} item
-	 * @return {String|Number}
-	 */
-	resolveIdForItem: function( type, item = null ) {
-		if ( item === null ) {
-			return null;
-		}
-
-		if ( _.isString( item ) || _.isNumber( item ) ) {
-			return item;
-		}
-
-		return item.id || item[ type.prototype.idAttribute ] || null;
-	},
-
-	/**
-	 * Find a specific model of a certain `type` in the store
-	 * @param type
-	 * @param {String|Number|Object|Backbone.Relational.Model} item
-	 */
-	find: function( type, item ) {
-		var id = this.resolveIdForItem( type, item ),
-			coll = this.getCollection( type );
-
-		// Because the found object could be of any of the type's superModel
-		// types, only return it if it's actually of the type asked for.
-		if ( coll ) {
-			var obj = coll.get( id );
-
-			if ( obj instanceof type ) {
-				return obj;
-			}
-		}
-
-		return null;
-	},
-
-	/**
-	 * Add a 'model' to its appropriate collection. Retain the original contents of 'model.collection'.
-	 * @param {Backbone.Relational.Model} model
-	 */
-	register: function( model ) {
-		var coll = this.getCollection( model );
-
-		if ( coll ) {
-			var modelColl = model.collection;
-			coll.add( model );
-			model.collection = modelColl;
-		}
-	},
-
-	/**
-	 * Check if the given model may use the given `id`
-	 * @param model
-	 * @param [id]
-	 */
-	checkId: function( model, id ) {
-		var coll = this.getCollection( model ),
-			duplicate = coll && coll.get( id );
-
-		if ( duplicate && model !== duplicate ) {
-			if ( config.showWarnings && console ) {
-				console.warn( 'Duplicate id! Old RelationalModel=%o, new RelationalModel=%o', duplicate, model );
-			}
-
-			throw new Error( "Cannot instantiate more than one Backbone.Relational.Model with the same id per type!" );
-		}
-	},
-
-	/**
-	 * Explicitly update a model's id in its store collection
-	 * @param {Backbone.Relational.Model} model
-	 */
-	update: function( model ) {
-		var coll = this.getCollection( model );
-
-		// Register a model if it isn't yet (which happens if it was created without an id).
-		if ( !coll.contains( model ) ) {
-			this.register( model );
-		}
-
-		// This triggers updating the lookup indices kept in a collection
-		coll._onModelEvent( 'change:' + model.idAttribute, model, coll );
-
-		// Trigger an event on model so related models (having the model's new id in their keyContents) can add it.
-		model.trigger( 'relational:change:id', model, coll );
-	},
-
-	/**
-	 * Unregister from the store: a specific model, a collection, or a model type.
-	 * @param {Backbone.Relational.Model|Backbone.Relational.Model.constructor|Backbone.Relational.Collection} type
-	 */
-	unregister: function( type ) {
-		let coll;
-		let models = _.clone( type.models );
-		if ( type.model ) {
-			coll = this.getCollection( type.model );
-		}
-		else {
-			coll = this.getCollection( type );
-			models = _.clone( coll.models );
-		}
-
-		if ( type instanceof Backbone.Model ) {
-			models = [ type ];
-		}
-
-		_.each( models, function( model ) {
-			this.stopListening( model );
-			_.invoke( model.getRelations(), 'stopListening' );
-		}, this );
-
-		// If we've unregistered an entire store collection, reset the collection (which is much faster).
-		// Otherwise, remove each model one by one.
-		if ( _.contains( this._collections, type ) ) {
-			coll.reset( [] );
-		}
-		else {
-			_.each( models, function( model ) {
-				if ( coll.get( model ) ) {
-					coll.remove( model );
-				}
-				else {
-					coll.trigger( 'relational:remove', model, coll );
-				}
-			}, this );
-		}
-	},
-
-	/**
-	 * Reset the `store` to it's original state. The `reverseRelations` are kept though, since attempting to
-	 * re-initialize these on models would lead to a large amount of warnings.
-	 */
-	reset: function() {
-		this.stopListening();
-
-		// Unregister each collection to remove event listeners
-		_.each( this._collections, function( coll ) {
-			this.unregister( coll );
-		}, this );
-
-		this._collections = [];
-		this._subModels = [];
-		this._modelScopes = [ window ];
-	}
-});
-
-module.store = new module.Store();
-
-/**
- * The main Relation class, from which 'HasOne' and 'HasMany' inherit. Internally, 'relational:<key>' events
- * are used to regulate addition and removal of models from relations.
- *
- * @param {Backbone.Relational.Model} [instance] Model that this relation is created for. If no model is supplied,
- *      Relation just tries to instantiate it's `reverseRelation` if specified, and bails out after that.
- * @param {Object} options
- * @param {string} options.key
- * @param {Backbone.Relational.Model.constructor} options.relatedModel
- * @param {Boolean|String} [options.includeInJSON=true] Serialize the given attribute for related model(s)' in toJSON, or just their ids.
- * @param {Boolean} [options.createModels=true] Create objects from the contents of keys if the object is not found in Backbone.store.
- * @param {Object} [options.reverseRelation] Specify a bi-directional relation. If provided, Relation will reciprocate
- *    the relation to the 'relatedModel'. Required and optional properties match 'options', except that it also needs
- *    {Backbone.Relation|String} type ('HasOne' or 'HasMany').
- * @param {Object} opts
- */
-module.Relation = BObject.extend( Semaphore ).extend({
-	instance: null,
-	key: null,
-	keyContents: null,
-	relatedModel: null,
-	relatedCollection: null,
-	reverseRelation: null,
-	related: null,
-
-	constructor( instance, options, opts ) {
-		this.instance = instance;
-		// Make sure 'options' is sane, and fill with defaults from subclasses and this object's prototype
-		options = _.isObject( options ) ? options : {};
-		this.reverseRelation = _.defaults( options.reverseRelation || {}, this.options.reverseRelation );
-		this.options = _.defaults( options, this.options, {
-			createModels: true,
-			includeInJSON: true,
-			isAutoRelation: false,
-			autoFetch: false,
-			parse: false
-		});
-
-		// TODO: handle this logic else where. it's crazy to have a parent depend on its children!!
-		this.reverseRelation.type = !_.isString( this.reverseRelation.type ) ? this.reverseRelation.type :
-			module[ this.reverseRelation.type ] || module.store.getObjectByName( this.reverseRelation.type );
-
-		this.key = this.options.key;
-		this.keySource = this.options.keySource || this.key;
-		this.keyDestination = this.options.keyDestination || this.keySource || this.key;
-
-		this.model = this.options.model || this.instance.constructor;
-
-		this.relatedModel = this.options.relatedModel;
-
-		if ( _.isUndefined( this.relatedModel ) ) {
-			this.relatedModel = this.model;
-		}
-
-		// if ( _.isFunction( this.relatedModel ) && !( this.relatedModel.prototype instanceof Backbone.Model ) ) {
-		// 	this.relatedModel = _.result( this, 'relatedModel' );
-		// }
-		if ( _.isString( this.relatedModel ) ) {
-			this.relatedModel = module.store.getObjectByName( this.relatedModel );
-		}
-
-		if ( !this.checkPreconditions() ) {
-			return;
-		}
-
-		// Add the reverse relation on 'relatedModel' to the store's reverseRelations
-		if ( !this.options.isAutoRelation && this.reverseRelation.type && this.reverseRelation.key ) {
-			module.store.addReverseRelation( _.defaults({
-					isAutoRelation: true,
-					model: this.relatedModel,
-					relatedModel: this.model,
-					reverseRelation: this.options // current relation is the 'reverseRelation' for its own reverseRelation
-				},
-				this.reverseRelation // Take further properties from this.reverseRelation (type, key, etc.)
-			) );
-		}
-
-		if ( instance ) {
-			var contentKey = this.keySource;
-			if ( contentKey !== this.key && _.isObject( this.instance.get( this.key ) ) ) {
-				contentKey = this.key;
-			}
-
-			this.setKeyContents( this.instance.get( contentKey ) );
-			this.relatedCollection = module.store.getCollection( this.relatedModel );
-
-			// Explicitly clear 'keySource', to prevent a leaky abstraction if 'keySource' differs from 'key'.
-			if ( this.keySource !== this.key ) {
-				delete this.instance.attributes[ this.keySource ];
-			}
-
-			// Add this Relation to instance._relations
-			this.instance._relations[ this.key ] = this;
-
-			this.initialize( opts );
-
-			if ( this.options.autoFetch ) {
-				this.instance.getAsync( this.key, _.isObject( this.options.autoFetch ) ? this.options.autoFetch : {} );
-			}
-
-			// When 'relatedModel' are created or destroyed, check if it affects this relation.
-			this.listenTo( this.instance, 'destroy', this.destroy )
-				.listenTo( this.relatedCollection, 'relational:add relational:change:id', this.tryAddRelated )
-				.listenTo( this.relatedCollection, 'relational:remove', this.removeRelated );
-		}
-	},
-
-	/**
-	 * Check several pre-conditions.
-	 * @return {Boolean} True if pre-conditions are satisfied, false if they're not.
-	 */
-	checkPreconditions: function() {
-		var i = this.instance,
-			k = this.key,
-			m = this.model,
-			rm = this.relatedModel,
-			warn = module.showWarnings && typeof console !== 'undefined';
-
-		if ( !m || !k || !rm ) {
-			warn && console.warn( 'Relation=%o: missing model, key or relatedModel (%o, %o, %o).', this, m, k, rm );
-			return false;
-		}
-		// Check if the type in 'model' inherits from Backbone.Relational.Model
-		// if ( !( m.prototype instanceof module.Model ) ) {
-		// 	warn && console.warn( 'Relation=%o: model does not inherit from Backbone.Relational.Model (%o).', this, i );
-		// 	return false;
-		// }
-		// Check if the type in 'relatedModel' inherits from Backbone.Relational.Model
-		// if ( !( rm.prototype instanceof module.Model ) ) {
-		// 	warn && console.warn( 'Relation=%o: relatedModel does not inherit from Backbone.Relational.Model (%o).', this, rm );
-		// 	return false;
-		// }
-		// Check if this is not a HasMany, and the reverse relation is HasMany as well
-		// TODO: handle this logic elsewhere!
-		if ( this instanceof module.HasMany && this.reverseRelation.type === module.HasMany ) {
-			warn && console.warn( 'Relation=%o: relation is a HasMany, and the reverseRelation is HasMany as well.', this );
-			return false;
-		}
-		// Check if we're not attempting to create a relationship on a `key` that's already used.
-		if ( i && _.keys( i._relations ).length ) {
-			var existing = _.find( i._relations, function( rel ) {
-				return rel.key === k;
-			}, this );
-
-			if ( existing ) {
-				warn && console.warn( 'Cannot create relation=%o on %o for model=%o: already taken by relation=%o.',
-					this, k, i, existing );
-				return false;
-			}
-		}
-
-		return true;
-	},
-
-	/**
-	 * Set the related model(s) for this relation
-	 * @param {Backbone.Model|module.Collection} related
-	 */
-	setRelated: function( related ) {
-		this.related = related;
-		this.instance.attributes[ this.key ] = related;
-	},
-
-	/**
-	 * Determine if a relation (on a different RelationalModel) is the reverse
-	 * relation of the current one.
-	 * @param {Backbone.Relation} relation
-	 * @return {Boolean}
-	 */
-	_isReverseRelation: function( relation ) {
-		return relation.instance instanceof this.relatedModel && this.reverseRelation.key === relation.key &&
-			this.key === relation.reverseRelation.key;
-	},
-
-	/**
-	 * Get the reverse relations (pointing back to 'this.key' on 'this.instance') for the currently related model(s).
-	 * @param {Backbone.Relational.Model} [model] Get the reverse relations for a specific model.
-	 *    If not specified, 'this.related' is used.
-	 * @return {Backbone.Relation[]}
-	 */
-	getReverseRelations: function( model ) {
-		var reverseRelations = [];
-		// Iterate over 'model', 'this.related.models' (if this.related is a module.Collection), or wrap 'this.related' in an array.
-		var models = !_.isUndefined( model ) ? [ model ] : this.related && ( this.related.models || [ this.related ] ),
-			relations = null,
-			relation = null;
-
-		for( var i = 0; i < ( models || [] ).length; i++ ) {
-			relations = models[ i ].getRelations() || [];
-
-			for( var j = 0; j < relations.length; j++ ) {
-				relation = relations[ j ];
-
-
-				if ( this._isReverseRelation( relation ) ) {
-					reverseRelations.push( relation );
-				}
-			}
-		}
-
-		return reverseRelations;
-	},
-
-	/**
-	 * When `this.instance` is destroyed, cleanup our relations.
-	 * Get reverse relation, call removeRelated on each.
-	 */
-	destroy: function() {
-		this.stopListening();
-
-		if ( this instanceof module.HasOne ) {
-			this.setRelated( null );
-		}
-		else if ( this instanceof module.HasMany ) {
-			this.setRelated( this._prepareCollection() );
-		}
-
-		_.each( this.getReverseRelations(), function( relation ) {
-			relation.removeRelated( this.instance );
-		}, this );
-	}
-});
+module.Relation = Relation;
 
 module.HasOne = module.Relation.extend({
 	options: {
@@ -1018,7 +411,7 @@ module.HasMany = module.Relation.extend({
  *  - 'remove:<key>' (model, related collection, options)
  *  - 'change:<key>' (model, related model or collection, options)
  */
-module.Model = Backbone.Model.extend(Semaphore).extend({
+module.Model = Backbone.Model.extend( Semaphore ).extend({
 	relations: null, // Relation descriptions on the prototype
 	_relations: null, // Relation instances
 	_isInitialized: false,
@@ -1732,5 +1125,8 @@ module.Model = Backbone.Model.extend(Semaphore).extend({
 		return child;
 	}
 });
+
+relationTypeStore.registerType( 'HasOne', module.HasOne );
+relationTypeStore.registerType( 'HasMany', module.HasMany );
 
 export default module;
