+++ src/relation-type-store.js
@@ -1,0 +1,3 @@
+import RelationTypeStore from './utils/relation-type-store';
+
+export default new RelationTypeStore();
