+++ src/store.js
@@ -1,0 +1,3 @@
+import Store from './utils/store';
+
+export default new Store();
